Add new option to compile digiKAm core with Qt6::Multimedia instead internal QtAV. This option is disabled by default for the moment, and this is only concern Qt6 > 6.5. As last log of Qt6.5.3 indicate a lots of bugs fixed, the QtMultimedia support needs again more work to be stabilized from Qt team.
https://code.qt.io/cgit/qt/qtreleasenotes.git/about/qt/6.5.3/release-note.md#qtmultimedia

I have tested the qt::Multimedia support with Qt 6.5.2, and i noted plenty of dysfunction, as video artefacts, video lock, video playback wrong size initialized at playing time, and more.

So Qt6Multimedia is so far not yet stable as desired to be used in production. Perhaps in few Qt6 release in the future, i hope.

IMPORTANT: compared to Qt5::Multimedia, the Qt6 version API is not yet complete. It's not yet possible to stream customized
QVideoFrame in QMediaRecorder, and we needs this to create video from images in Video Slideshow plugin.