Save rejected face data
Save rejected face tags and exclude the tags from being used during face classification (matching).

This MR does not include code to write the rejected face tags to the item metadata.  That will be a different MR in the next several days.

This MR does not include the ability to clear the reject face data. That will be a different MR in the next several days.

BUG: 502219
BUG: 444394
BUG: 432207
BUG: 415783
BUG: 502924