Add new method to run digiKam or Showfoto as attached to a remote GDB instance.
This is a completly different approach than previous one for the AppImage bundles where the application was
run directly in GDB within the AppRun bash script used by the AppImage. This method introduce binary incompatibilities
with GDB and the application which are compiled on different Linux versions.

A new env.var DIGIKAM_REMOTE_DEBUG must be turn ON to before to run the application. A 30 s loop let's time to
the user to start the debugger instance to capture the application process before to start initial scan.

For the AppImage, due to the sandboxing and the security issue with the remote debugging, the bundle contents
must be previously extracted locally and GDB must be started with sudo.

This mechanism is inspired from https://lavaforge.org/preserve-emulation/pcsx2/wiki/15-Using-GDB-with-Linux-AppImages

TODO: update the contrib page from the web site to explain all the debugging steps.