FaceRegion: fix crash if rectangle coordinates are invalid.
A crash can be caused by an integer overflow in the QRect constructor, triggered by invalid XML data in the face region descriptor.
Use case of data values extracted from the database: <rect x="-2147483648" y="-2147483648" width="-2147483648" height="-2147483648"/>

The values -2147483648 are the minimum possible values for a 32-bit signed integer (INT_MIN),
which, when used in arithmetic operations, can cause overflows. The QRect constructor internally
uses QCheckedInt to detect overflows, and it aborts the program when it encounters one.